% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{bs4Alert}
\alias{bs4Alert}
\title{Create alert in UI}
\usage{
bs4Alert(..., color = "primary", dismissable = FALSE)
}
\arguments{
\item{...}{One or more HTML elements to show in the alert}

\item{color}{The color of the alert}

\item{dismissable}{Should the user be allowed to dismiss the alert}
}
\description{
Create a Bootstrap alert box that optionally can be dismissed by the user.
}
