% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{bs4Embed}
\alias{bs4Embed}
\title{Create responsive embed}
\usage{
bs4Embed(src, type = c("iframe", "video"), ratio = c(16, 9))
}
\arguments{
\item{src}{The URL to the source that will be embedded}

\item{type}{The type of source (usually an `iframe`)}

\item{ratio}{A vector of lenght 2 with the aspect ratio for the embedded content.
See details}
}
\description{
Embed content from and external source and make the embed responsive.
}
\details{
Supported aspect ratios are 16 by 9 (default), 4 by 3, 21 by 9 and 1 by 1.
}
