% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs4Lib}
\alias{bs4Lib}
\title{Bootstrap Libraries}
\usage{
bs4Lib(theme = NULL)
}
\arguments{
\item{theme}{Alternate Bootstrap 4 stylesheet.}
}
\description{
This function returns a set of web dependencies necessary for using Bootstrap
components in a web page
}
