% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{bs4Modal}
\alias{bs4Modal}
\alias{modalDialog}
\title{Create a modal dialog}
\usage{
bs4Modal(
  ...,
  title = NULL,
  valign = FALSE,
  size = "m",
  footer = bs4ModalButton(),
  easyClose = FALSE,
  fade = TRUE
)

modalDialog(
  ...,
  title = NULL,
  valign = FALSE,
  size = "m",
  footer = bs4ModalButton(),
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{...}{UI elements for the body of the modal dialog box.}

\item{title}{An optional title for the modal.}

\item{valign}{Boolean. If \code{TRUE} the modal is vertically centered}

\item{size}{The size of the modal. One of \code{s}, \code{m} (default), \code{l} or \code{xl}}

\item{footer}{UI for the footer of the modal}

\item{easyClose}{Boolean. If \code{TRUE} the modal can be dismissed by clicking outside
the dialog box}
}
\description{
Create a modal dialog
}
