% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{bs4ModalButton}
\alias{bs4ModalButton}
\alias{modalButton}
\title{Create a modal button}
\usage{
bs4ModalButton(
  label = "Close",
  icon = NULL,
  color = "secondary",
  size = c("normal", "sm", "lg")
)

modalButton(
  label = "Close",
  icon = NULL,
  color = "secondary",
  size = c("normal", "sm", "lg")
)
}
\arguments{
\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[=createIcon]{icon}} to appear next to the label}

\item{color}{The color of the button. Must be a valid Bootstrap 4 color.}

\item{size}{The size of the button. Use \code{sm} for small and \code{lg} for
large buttons.}
}
\description{
Create a button that will dismiss a modal dialog.
}
\seealso{
\code{\link{bs4ModalDialog}}
}
