% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs4Page}
\alias{bs4Page}
\title{Bootstrap Page}
\usage{
bs4Page(..., title = NULL, theme = NULL)
}
\arguments{
\item{...}{A parameter}

\item{title}{The title for the page}

\item{theme}{A parameter}
}
\description{
Bootstrap Page
}
