% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{cardGroup}
\alias{cardGroup}
\title{Create a group of two or more \code{\link{dashboardCard}}s}
\usage{
cardGroup(..., type = c("group", "deck"), class = NULL)
}
\arguments{
\item{...}{Two or more \code{\link{dashboardCard}}s to go into the group.}

\item{type}{The type of card group. Can be either \code{group} or \code{deck}.}

\item{class}{Additional CSS classes to apply to the div tag}
}
\description{
Create a group of two or more \code{\link{dashboardCard}}s. Two types of groups are
supported; \code{group} and \code{deck}. See details.
}
\details{
The cards the the group will be of equal height and fill the full width on big screen devices.
The cards are responsive. When the \code{type} is set to \code{group}, the cards are grouped
together with no space between them. Headers and footers will automatically line up. When the
\code{type} is set to \code{deck} the cards will be set to equal width and height, but will
show as individual cards.
}
\seealso{
\code{\link{dashboardCard}}
}
