% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{cardNav}
\alias{cardNav}
\title{Card with navigation}
\usage{
cardNav(navId, ...)
}
\arguments{
\item{navId}{The ID of the card}

\item{...}{One or more nav items of type \code{\link{cardNavItem}}}
}
\description{
Create a card with several \code{\link{cardNavItem}}s the user can navigate to. See details.
}
\examples{
# Build a card with three navigatable items
cardNav(
  navId = 'cardnav',
  cardNavItem(cardId = 'first', title = 'First', 'Content of first tab'),
  cardNavItem(cardId = 'second', title = 'Second', 'Content of second tab')
  cardNavItem(cardId = 'third', title = 'Third', 'Content of third tab')
)

}
