% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{cardNavItem}
\alias{cardNavItem}
\title{Item for card with navigation}
\usage{
cardNavItem(cardId, title = cardId, ...)
}
\arguments{
\item{cardId}{The ID for the card body}

\item{...}{UT elements to include within the card body}
}
\description{
Create a card item that can be included in a \code{\link{cardNav}}
}
\seealso{
\code{\link{cardNav}}
}
