% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{dashboardPanel}
\alias{dashboardPanel}
\title{Dashboard panel}
\usage{
dashboardPanel(title, ..., id = title, value = title, icon = "dashboard")
}
\arguments{
\item{title}{The title for the dashboard}

\item{...}{The UI elements for the dashboard. Can be any Shiny input or output, or a
\code{dashboardFilter}.}

\item{id}{The ID for the dashboard}

\item{value}{The value for the dashboard}

\item{icon}{A string giving the icon for the dashboard. Can be any icon from the OpenIconic
library. See details.}
}
\description{
Create a dashboard panel.
}
\details{
This function creates a dashboard panel for outputting results from R functions.
  Icons can be:
  * \code{dashboard}
  * \code{bar-chart}
  * \code{graph}
  * \code{map-marker}
  * \code{people}
  * \code{pulse}
  * \code{list}
  * \code{bolt}
  * \code{dollar}
  * \code{cog}
}
