% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{downloadButton}
\alias{downloadButton}
\title{Download button}
\usage{
downloadButton(outputId, label = "Download", color = "primary",
  outline = FALSE, class = NULL, size = c("normal", "sm", "lg"), ...)
}
\arguments{
\item{outputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button or link–usually a text label, but you
could also use any other HTML, like an image.}

\item{color}{Alternate Bootstrap 4 stylesheet.}

\item{outline}{Should the button be an outline button? Default value FALSE.}

\item{class}{Optional CSS classes.}

\item{size}{The size of the button.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Download button
}
