% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dropdownInput}
\alias{dropdownInput}
\alias{dropdownMenu}
\title{Create a dropdown menu input}
\usage{
dropdownInput(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  icon = NULL,
  color = "primary",
  outline = FALSE,
  size = c("normal", "sm", "lg"),
  direction = c("down", "right", "up", "left")
)

dropdownMenu(...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The text on the dropdown menu button.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{multiple}{If \code{TRUE}, multiple items can be selected}

\item{icon}{An optional \code{\link[=createIcon]{icon}} to appear next to the label}

\item{color}{The color of the button. Must be a valid Bootstrap 4 color.}

\item{outline}{If \code{TRUE}, background color will be removed from the button.
Defaults to \code{FALSE}.}

\item{size}{The size of the button. Use \code{sm} for small and \code{lg} for
large buttons.}

\item{direction}{The direction of the menu. Defaults to \code{'down'}.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Creates a dropdown menu input.
}
