% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dropdownMenu}
\alias{dropdownMenu}
\title{Create a dropdown menu for inputs}
\usage{
dropdownMenu(inputId, label, choices, selected = NULL, multiple = FALSE,
  icon = NULL, color = c("primary", "secondary", "success", "danger",
  "warning", "info", "light", "dark"), outline = FALSE, size = c("normal",
  "sm", "lg"), direction = c("down", "right", "up", "left"))
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The text on the dropdown menu button.}

\item{choices}{List of values to select from.}

\item{selected}{The initially selected value.}

\item{multiple}{Is selection of multiple items allowed?}

\item{icon}{Optional icon shown before the label.}

\item{color}{A character string giving the color of the search button.}

\item{outline}{Should the button be an outline button? Default value \code{TRUE}.}

\item{size}{A character string giving the size of the input. Valid options are \code{normal},
\code{sm} and \code{lg}. Defaults to \code{normal}}

\item{direction}{The direction of the menu. Default value \code{'up'}.}
}
\description{
Create a dropdown menu for inputs
}
