% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dropdownSelect}
\alias{dropdownSelect}
\title{Dropdown select}
\usage{
dropdownSelect(inputId, label, color = "", icon = NULL, width = NULL, ...,
  choices = c(), selected = c(), multiple = FALSE)
}
\description{
Dropdown select
}
