% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{flexRow}
\alias{flexRow}
\title{Create flexible row}
\usage{
flexRow(...)
}
\arguments{
\item{...}{The elements to include in the row}
}
\description{
Create a flex row using Bootstrap classes. For more information, see
  the \href{https://getbootstrap.com/docs/4.6/utilities/flex/}{Bootstrap documentation}
}
