% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{footerContent}
\alias{footerContent}
\title{Add content to page footer}
\usage{
footerContent(
  ...,
  color = "light",
  wrapper = TRUE,
  center = TRUE,
  small = FALSE
)
}
\arguments{
\item{...}{The text to be included}

\item{color}{The background color for the text. See details}

\item{wrapper}{If \code{TRUE} the text will be wrapped in a paragraph tag}

\item{center}{If \code{TRUE} the text will be centered}

\item{small}{If \code{TRUE} the text will be smaller than the body text}
}
\description{
Add a text message below the content area on \code{\link{saiPage}}
}
\details{
Color can be either a valid bootstrap color, or a hexadecimal string
of the form "\code{#rrggbb}". See \code{\link[grDevices]{rgb}}.
}
