% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{fullpageLayout}
\alias{fullpageLayout}
\title{Full page layout}
\usage{
fullpageLayout(...)
}
\arguments{
\item{...}{UI elements to include in the layout}
}
\description{
Create a full page (fuild) layout. This function should be used together with
\code{\link{saiPage}}.
}
