% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{helpText}
\alias{helpText}
\title{Create a help text element}
\usage{
helpText(..., small = FALSE)
}
\arguments{
\item{...}{One or more help text strings (or other inline HTML elements)}

\item{small}{Should the text be smaller?}
}
\value{
A help text element that can be added to a UI definition.
}
\description{
Create a help text element
}
