% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{saiDashboard}
\alias{saiDashboard}
\title{Dashboard page}
\usage{
saiDashboard(title, ..., color = "dark", brand = title,
  windowTitle = title)
}
\arguments{
\item{title}{Title}

\item{...}{Elements}

\item{color}{A string indicating the color of the top navigation}

\item{brand}{A string indicating the brand for the page}

\item{windowTitle}{The title that should be displayed by the browser window.}
}
\description{
Dashboard page
}
