% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{saiMain}
\alias{saiMain}
\title{Main content}
\usage{
saiMain(..., width = 8)
}
\arguments{
\item{...}{Output elements to include in the main panel}

\item{width}{The width of the main panel. For fluid layouts this is out of 12 total units;
for fixed layouts it is out of whatever the width of the sidebar's parent column is.}
}
\description{
Create a main panel containing output elements
}
