% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{saiMenu}
\alias{saiMenu}
\title{SAI menu}
\usage{
saiMenu(..., width = 4)
}
\arguments{
\item{...}{UI elements to include on the sidebar}

\item{width}{The width of the sidebar. For fluid layouts this is out of 12 total units;
for fixed layouts it is out of whatever the width of the sidebar's parent column is.}
}
\description{
Create a sidebar panel containing input controls.
}
