% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{saiTab}
\alias{saiTab}
\alias{tabPanel}
\title{Create a tab panel}
\usage{
saiTab(title, ..., value = title, class = NULL, icon = NULL, hidden = FALSE)

tabPanel(title, ..., value = title, class = NULL, icon = NULL, hidden = FALSE)
}
\arguments{
\item{title}{Display title for tab}

\item{...}{UI elements to include within the tab}

\item{value}{The value that should be sent when \code{tabsetPanel} reports
that this tab is selected. If omitted and \code{tabsetPanel} has an
\code{id}, then the title will be used.}

\item{class}{Additional classes to add to the tab.}

\item{icon}{An optional \code{\link[=createIcon]{icon}} to appear on the tab.}

\item{hidden}{Boolean value. Should the tab be hidden from the menu? Defalts to \code{FALSE}}
}
\description{
Create a tab panel that can be included within a \code{tabsetPanel}. Overrides default Shiny
  \code{tabPanel}.
}
