% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{saiTabset}
\alias{saiTabset}
\alias{tabsetPanel}
\title{Create a tabset panel}
\usage{
saiTabset(..., id = NULL, selected = NULL, type = c("tabs", "pills"))

tabsetPanel(..., id = NULL, selected = NULL, type = c("tabs", "pills"))
}
\arguments{
\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the tabset}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{type}{\describe{
\item{\code{"tabs"}}{Standard tab look}
\item{\code{"pills"}}{Selected tabs use the background fill color}
\item{\code{"hidden"}}{Hides the selectable tabs. Use \code{type = "hidden"} in
conjunction with \code{\link[shiny:tabPanel]{tabPanelBody()}} and \code{\link[shiny:updateTabsetPanel]{updateTabsetPanel()}} to control the
active tab via other input controls. (See example below)}
}}
}
\value{
A tabset that can be passed to \code{\link{mainPanel}}
}
\description{
Create a tabset that contains \code{\link{tabPanel}} elements. Tabsets are
useful for dividing output into multiple independently viewable sections.
}
\examples{
# Show a tabset that includes a plot, summary, and
# table view of the generated distribution
mainPanel(
  tabsetPanel(
    tabPanel("Plot", plotOutput("plot")),
    tabPanel("Summary", verbatimTextOutput("summary")),
    tabPanel("Table", tableOutput("table"))
  )
)
}
\seealso{
\code{\link{tabPanel}}, \code{\link{updateTabsetPanel}}
}
