% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{searchboxInput}
\alias{searchboxInput}
\title{Searchbox input}
\usage{
searchboxInput(inputId, value = "", button = "Search", placeholder = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{button}{A character string to display as the text for the search button.}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the
control.}
}
\description{
Searchbox input
}
