% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toasts.R
\name{sendToast}
\alias{sendToast}
\title{Send toast notification to the UI}
\usage{
sendToast(
  title,
  message,
  session = shiny::getDefaultReactiveDomain(),
  autohide = TRUE,
  delay = 5000
)
}
\arguments{
\item{title}{A character vector with the title of the notification}

\item{message}{A character vector or a tag object with the content of the notification}

\item{session}{The Shiny session object}

\item{autohide}{If \code{TRUE}, the notification will automatically be hidden after
a set delay}

\item{delay}{A numeric value giving the delay in milliseconds before hiding the toast}
}
\description{
Sends a toast notification to the user. Requires that a container for toast
notifications has been enabled. See details
}
