% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sidebarLayout}
\alias{sidebarLayout}
\title{Sidebar layout}
\usage{
sidebarLayout(menu, main, position = c("left", "right"), fluid = TRUE)
}
\arguments{
\item{menu}{The \code{\link{saiMenu}} containing input controls}

\item{main}{The \code{\link{saiMain}} containing outputs}

\item{position}{The position of the menu relative to the main area}

\item{fluid}{\code{TRUE} to use a fluid layout (100% width on all devices), or \code{FALSE}
to use a responsive layout. Defaults to \code{TRUE}}
}
\description{
Sidebar layout
}
