% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{singlePage}
\alias{singlePage}
\title{Create a simple Bootstrap page}
\usage{
singlePage(title, ..., theme = NULL, lang = "en")
}
\arguments{
\item{title}{The title for the page.}

\item{...}{The UI elements of the page.}

\item{theme}{A character string or a \code{\link[htmltools]{htmlDependency}}-object
with an alternate Bootstrap 4 stylesheet.}

\item{lang}{The language of the page that is added to the top level \code{html} element.}
}
\description{
Create a simple Bootstrap page
}
