% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{slicerInput}
\alias{slicerInput}
\title{Slicer Input}
\usage{
slicerInput(
  inputId,
  label,
  choices,
  selected = NULL,
  color = "primary",
  outline = TRUE,
  multiple = FALSE
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button or link.}

\item{choices}{A list of choices to select from. Each choice will be displayed as a button.}

\item{selected}{A single value or a vector of values that will be initially selected.}

\item{color}{Alternate Bootstrap 4 stylesheet.}

\item{outline}{Should the button be an outline button? Default value \code{TRUE}.}

\item{multiple}{Is selection of multiple items allowed?}
}
\value{
A list control that can be added to a UI definition.
}
\description{
Create a slicer -- a set of buttons that can act as a filter for a dataset.
}
