% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{switchInput}
\alias{switchInput}
\title{Switch input}
\usage{
switchInput(inputId, label, value = FALSE)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}
}
\value{
A switch input that can be added to a UI definition
}
\description{
Create a switch input that can be used to specify logical values similar to a checkbox
}
\seealso{
\code{\link[shiny]{checkboxInput}}
}
