% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{tabPanel}
\alias{tabPanel}
\alias{saiTab}
\title{Create a tab panel}
\usage{
tabPanel(title, ..., value = title, icon = NULL)

saiTab(title, ..., value = title, icon = NULL)
}
\arguments{
\item{title}{Display title for tab.}

\item{...}{UI elements to include within the tab.}

\item{value}{The value that should be sent when \code{tabsetPanel} reports that this tab is selected.}

\item{icon}{Optional icon to appear on the tab.}
}
\description{
Create a tab panel that can be included within a \code{tabsetPanel}. Overrides default Shiny
  \code{tabPanel}.
}
