% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toasts.R
\name{toastWrapper}
\alias{toastWrapper}
\title{Add a container for sending toast notifications}
\usage{
toastWrapper()
}
\description{
Manually insert a toast container. For \code{\link{saiPage}} and \code{\link{saiDashboard}},
use the \code{notifications} argument instead. If you want to override the positioning
of toast notifications, you need to set this argument to \code{FALSE}.
}
