% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{toggleButton}
\alias{toggleButton}
\title{Toggle button}
\usage{
toggleButton(inputId, label, color = "primary", outline = FALSE,
  size = c("normal", "sm", "lg"), active = FALSE)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button or link.}

\item{color}{Alternate Bootstrap 4 stylesheet.}

\item{outline}{Should the button be an outline button? Default value \code{FALSE}.}

\item{size}{The size of the button.}

\item{active}{Should the button be active on initalization? Default value \code{FALSE}.}
}
\description{
Creates a button that can be toggled on and off. Returns a \code{TRUE} value when the
button is pressed and a \code{FALSE} value when the button is not pressed. Defaults to
not pressed (not active).
}
