% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{updateSlicerInput}
\alias{updateSlicerInput}
\title{Update slicer input}
\usage{
updateSlicerInput(session, inputId, value = NULL, change = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{value}{Optional. The new value of the button.}

\item{change}{Optional. If the current value should be changed.}
}
\description{
Change the state of the slicer input on the client.
}
\details{
If both \code{value} and \code{change} is given, \code{value} is given priority.
If neither \code{value} nor \code{change} is given, \code{change} is assumed.
}
