% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{updateToggleButton}
\alias{updateToggleButton}
\title{Update Toggle Button}
\usage{
updateToggleButton(session, inputId, value = NULL, change = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{value}{Optional. The new value of the button.}

\item{change}{Optional. If the current value should be changed.}

\item{inputIde}{The id of the input object.}
}
\description{
Change the state of the toggle button on the client.
}
\details{
If both \code{value} and \code{change} is given, \code{value} is given priority.
If neither \code{value} nor \code{change} is given, \code{change} is assumed.
}
