% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{actionButton}
\alias{actionButton}
\title{Action button}
\usage{
actionButton(inputId, label, color = "primary", icon = NULL, width = NULL,
  ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The contents of the button or link–usually a text label, but you
could also use any other HTML, like an image.}

\item{color}{Alternate Bootstrap 4 stylesheet.}

\item{icon}{An optional icon to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'}}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Action button
}
