% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{bs4Alert}
\alias{bs4Alert}
\title{Create alert in UI}
\usage{
bs4Alert(..., color = "primary", icon = NULL, dismissable = FALSE)
}
\arguments{
\item{...}{One or more HTML elements to show in the alert}

\item{color}{The color of the alert}

\item{icon}{An optional \code{\link[=createIcon]{icon}} to appear next to the text}

\item{dismissable}{If \code{TRUE} the user can dismiss the alert. Defaults to
\code{FALSE}.}
}
\description{
Create a Bootstrap alert box that optionally can be dismissed by the user.
}
\details{
This creates a Bootstrap 4 type alert that can be shown in the UI. The alert can
optionally be dismissed by the user, by setting the argument \code{dismissable} to
\code{TRUE}.

Alerts can be placed anywhere in the UI. They are useful for highlighting important
information to the user. If you want to send a message to the user,
\code{\link[=sendToast]{toast notifications}} might be a better alternative.
}
\examples{
bs4Alert('This is important information', color = 'warning', icon = 'warning')

}
\seealso{
\code{\link{sendToast}}
}
