% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs4Lib}
\alias{bs4Lib}
\alias{saiLib}
\title{Bootstrap Libraries}
\usage{
bs4Lib(theme = TRUE, deps = NULL)

saiLib(theme = TRUE, deps = NULL)
}
\arguments{
\item{theme}{If \code{TRUE} the default Bootstrap theme will be added.}

\item{deps}{Additional dependencies to add to the page.}
}
\description{
This function returns a set of web dependencies necessary for using Bootstrap
components in a web page
}
