% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{bs4ModalButton}
\alias{bs4ModalButton}
\title{Add modal button}
\usage{
bs4ModalButton(text = "Close", color = "secondary")
}
\arguments{
\item{text}{The text of the close button.}
}
\description{
Add modal button
}
