% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bs4Page}
\alias{bs4Page}
\title{Bootstrap Page}
\usage{
bs4Page(..., title = NULL, theme = NULL, deps = NULL, lang = "en",
  dir = "ltr")
}
\arguments{
\item{...}{The UI elements of the page.}

\item{title}{The title for the page.}

\item{theme}{Alternate Bootstrap 4 stylesheet.}

\item{deps}{Additional dependencies to add to the page.}

\item{lang}{The language of the page that is added to the top level \code{html} element.}

\item{dir}{Indicate the directionality of text in the \code{html} page}
}
\description{
Bootstrap Page
}
