% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{downloadButton}
\alias{downloadButton}
\title{Create a download button or link}
\usage{
downloadButton(outputId, label = "Download", color = "primary",
  outline = FALSE, class = NULL, size = c("normal", "sm", "lg"), ...)
}
\arguments{
\item{outputId}{The name of the output slot that the \code{downloadHandler}
is assigned to.}

\item{label}{The label that should appear on the button.}

\item{color}{Alternate Bootstrap 4 stylesheet.}

\item{outline}{Should the button be an outline button? Default value FALSE.}

\item{class}{Additional CSS classes to apply to the tag, if any.}

\item{size}{The size of the button.}

\item{...}{Other arguments to pass to the container tag function.}
}
\description{
Use these functions to create a download button or link; when clicked, it
will initiate a browser download. The filename and contents are specified by
the corresponding \code{\link{downloadHandler}} defined in the server
function.
}
\examples{
\dontrun{
# In server.R:
output$downloadData <- downloadHandler(
  filename = function() {
    paste('data-', Sys.Date(), '.csv', sep='')
  },
  content = function(con) {
    write.csv(data, con)
  }
)

# In ui.R:
downloadLink('downloadData', 'Download')
}
}
\seealso{
\code{\link{downloadHandler}}
}
