% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{downloadButton}
\alias{downloadButton}
\title{Create a download button or link}
\usage{
downloadButton(
  outputId,
  label = "Download",
  color = "primary",
  outline = FALSE,
  class = NULL,
  size = c("normal", "sm", "lg"),
  ...
)
}
\arguments{
\item{outputId}{The name of the output slot that the \code{downloadHandler}
is assigned to.}

\item{label}{The label that should appear on the button.}

\item{color}{The color of the button. Must be a valid Bootstrap 4 color.}

\item{outline}{If \code{TRUE}, background color will be removed from the button.
Defaults to \code{FALSE}.}

\item{class}{Additional CSS classes to apply to the tag, if any.}

\item{size}{The size of the button. Use \code{sm} for small and \code{lg} for
large buttons.}

\item{...}{Other arguments to pass to the container tag function.}
}
\description{
Use these functions to create a download button or link; when clicked, it
will initiate a browser download. The filename and contents are specified by
the corresponding \code{\link[shiny:downloadHandler]{downloadHandler()}} defined in the server
function.
}
\examples{
\dontrun{
ui <- fluidPage(
  downloadButton("downloadData", "Download")
)

server <- function(input, output) {
  # Our dataset
  data <- mtcars

  output$downloadData <- downloadHandler(
    filename = function() {
      paste("data-", Sys.Date(), ".csv", sep="")
    },
    content = function(file) {
      write.csv(data, file)
    }
  )
}

shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:downloadHandler]{downloadHandler()}}
}
