% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{footerContent}
\alias{footerContent}
\title{Add content to page footer}
\usage{
footerContent(..., color = "light", center = TRUE, small = FALSE)
}
\arguments{
\item{...}{The text to include in the footer}

\item{color}{The background color for the text. Must be a valid Bootstrap 4 color}

\item{center}{Boolean. If \code{TRUE} the text will be centered}

\item{small}{Boolean. If \code{TRUE} the text will be smaller than the body text}
}
\description{
Add a text message below the content area on \code{\link{saiPage}}
}
