% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{footerContent}
\alias{footerContent}
\title{Add content to page footer}
\usage{
footerContent(..., color = "light")
}
\arguments{
\item{...}{The text to include in the footer}

\item{color}{The background color for the text. Must be a valid Bootstrap 4 color}
}
\description{
Add a text message below the content area on \code{\link{saiPage}}
}
