% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{headerContent}
\alias{headerContent}
\title{Add content to page header}
\usage{
headerContent(..., color = "primary")
}
\arguments{
\item{...}{The text to include in the header}

\item{color}{The background color for the text. Must be a valid Bootstrap 4 color}
}
\description{
Add a text message above the content area on \code{\link{saiPage}}
}
