% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{saiDashboard}
\alias{saiDashboard}
\title{Dashboard page}
\usage{
saiDashboard(title, ..., selected = NULL, color = "dark",
  brand = title, windowTitle = title, header = NULL, footer = NULL,
  lang = "en", dir = "ltr")
}
\arguments{
\item{title}{The title for the page.}

\item{...}{The UI elements of the page.}

\item{selected}{The \code{value} of the panel that should be selected by default.
If \code{NULL}, the first panel will be selected.}

\item{color}{A string indicating the color of the top navigation}

\item{brand}{A string indicating the brand for the page}

\item{windowTitle}{The title that should be displayed by the browser window.}

\item{header}{Tag or list of tags to display on top of page.}

\item{footer}{Tag or list of tags to display on the botton of the page.}

\item{lang}{The language of the page that is added to the top level \code{html} element.}

\item{dir}{Indicate the directionality of text in the \code{html} page}
}
\description{
Create a dashboard page that can be used with \code{\link{dashboardPanel}}s.
}
\seealso{
\code{\link{dashboardPanel}}
}
