% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{saiPage}
\alias{saiPage}
\title{SAI page element}
\usage{
saiPage(
  title,
  ...,
  id = NULL,
  selected = NULL,
  header = NULL,
  footer = NULL,
  theme = NULL,
  color = "primary",
  notifications = TRUE,
  windowTitle = title,
  lang = "en",
  dir = "ltr"
)
}
\arguments{
\item{title}{The title to display in the navbar.}

\item{...}{The UI elements of the page. Top level elements should be \code{\link{tabPanel}}.}

\item{id}{The page ID}

\item{selected}{The tab that is initially selected}

\item{header}{Tag or list of tags to display as a common header above all tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all tabPanels.}

\item{theme}{A character string or a \code{\link[htmltools]{htmlDependency}}-object
with an alternate Bootstrap 4 stylesheet.}

\item{color}{Color for the navbar. Supports all Bootstrap 4 colors.}

\item{notifications}{If \code{TRUE} (default), toast notifications are enabled}

\item{windowTitle}{The title that should be displayed by the browser window.}

\item{lang}{The language of the page that is added to the top level \code{html} element.}

\item{dir}{Indicate the directionality of text in the \code{html} page}
}
\description{
This functions builds the actual HTML page.
}
