% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{saiPage}
\alias{saiPage}
\title{SAI page element}
\usage{
saiPage(title, ..., id = NULL, selected = NULL, position = c(),
  header = NULL, footer = NULL, inverse = FALSE, fluid = TRUE,
  theme = NULL, windowTitle = title)
}
\description{
This functions builds the actual HTML page.
}
