% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{searchboxInput}
\alias{searchboxInput}
\title{Searchbox input}
\usage{
searchboxInput(inputId, value = "", placeholder = NULL, button = "Search",
  color = c("success", "primary", "secondary", "danger", "warning", "light",
  "dark"), size = c("normal", "sm", "lg"))
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{A character string with the default value of the search box}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the
control.}

\item{button}{A character string to display as the text for the search button.}

\item{color}{A character string giving the color of the search button.}

\item{size}{A character string giving the size of the input. Valid options are \code{normal},
\code{sm} and \code{lg}. Defaults to \code{normal}}
}
\description{
Searchbox input
}
