% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{searchboxInput}
\alias{searchboxInput}
\title{Create a searchbox input}
\usage{
searchboxInput(
  inputId,
  value = "",
  placeholder = NULL,
  button = NULL,
  icon = NULL,
  color = "primary",
  outline = FALSE,
  size = c("normal", "sm", "lg"),
  searchAsYouType = FALSE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{value}{A character string with the default value of the search box}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the
control.}

\item{button}{A character string to display as the text for the search button.}

\item{icon}{Name of an icon from the Open Iconic library instead of text.}

\item{color}{A character string giving the color of the search button.}

\item{outline}{Should the button be an outline button? Default value FALSE.}

\item{size}{A character string giving the size of the input. Valid options are \code{normal},
\code{sm} and \code{lg}. Defaults to \code{normal}}

\item{searchAsYouType}{If \code{TRUE}, the value will be continously updated to
Shiny. See details.}
}
\description{
Create an input control for text that behaves as a search field.
}
\details{
If \code{searchAsYouType} is set to \code{TRUE}, values will be sent to Shiny as
the user types (with a slight delay to account for continous typing). If the input
invalidates an expensive reactive function on the server, \code{searchAsYouType}
should be set to \code{FALSE}. This is also the default behaviour. When set to
\code{FALSE}, the value will not be updated before the user clicks on the button or
hits the enter key.
}
