% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{singleLayout}
\alias{singleLayout}
\title{Single column layout}
\usage{
singleLayout(..., max.width = FALSE)
}
\arguments{
\item{...}{UI elements to include in the layout}

\item{max.width}{The maximum width of the container i pixels. Default 1080.}
}
\description{
Single column layout
}
