% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{singleLayout}
\alias{singleLayout}
\title{Single column layout}
\usage{
singleLayout(..., fluid = FALSE)
}
\arguments{
\item{...}{UI elements to include in the layout}

\item{fluid}{If \code{TRUE} use a fluid layout (100% width on all devices). If \code{FALSE}
use a responsive layout. Defalts to \code{FALSE}}
}
\description{
Create a single coloumn responsive layout. This function should be used together with
\code{\link{saiPage}}.
}
\examples{
# Simple "Hello world" example
saiPage(
  title = "Demo page",
  tabPanel(
    singleLayout(h1("Hello world"))
  )
)
}
